+++ packages/passport/src/app.ts
@@ -17,13 +17,13 @@ import {
   AUTHORIZATION_CODE_FLOW,
   INVALID_ACCESS_SELF_INSTANCE_ERROR,
   INVALID_ACCESS_SERVER_ENV_ERROR,
+  INVALID_AUTHORIZATION_CODE_FLOW,
   INVALID_AUTH_SERVER_DOMAIN,
   INVALID_CACHE_LOCATION,
   INVALID_TOKEN_ROTATION,
   NOT_FOUND_ACCESS_TOKEN,
   NOT_FOUND_AUTH_IDENTIIFER,
   NOT_FOUND_ID_TOKEN,
-  NOT_FOUND_QUERY_PARAMS_ERROR,
   NOT_FOUND_REFRESH_TOKEN,
   NOT_FOUND_REFRESH_TOKEN_EXPIRES,
   NOT_FOUND_VALID_CLIENT_ID,
@@ -117,8 +117,11 @@ export class Passport {
       redirect_uri = redirect_uri.split('?code')[0];
     }
 
+    const useWorker = options.useWorker ?? false;
+
     this.#options = {
       ...options,
+      useWorker,
       authorizationOptions: {
         ...this.#defaultOptions.authorizationOptions,
         ...options.authorizationOptions,
@@ -143,8 +146,6 @@ export class Passport {
     const baseDomain = getDomain(options.domain);
 
     this.#authUrl = `${baseDomain}`;
-
-    const useWorker = options.useWorker ?? true;
 
     if (window.SharedWorker && useWorker) {
       this.#authWorker = new AuthWorker();
@@ -182,8 +183,8 @@ export class Passport {
     return auth_entry?.token;
   }
 
-  get #supportWorkerThread() {
-    return this.#authWorker instanceof SharedWorker;
+  get #supportAuthWorker() {
+    return Boolean(this.#authWorker) && this.#options.useWorker;
   }
 
   #getUrl<T extends EntireAuthorizationOptions | EntireAccessTokenOptions>(req: string, options: T) {
@@ -258,7 +259,7 @@ export class Passport {
 
   async #reconcileAuthorizationCodeFlow(error: any, onLoad: OnLoad) {
     if (this.checkAuthenticationError(error)) {
-      this.#transactionManager.remove();
+      this.#initializeTransaction();
 
       if (onLoad === 'check-sso') {
         this.#cacheManager.remove();
@@ -287,7 +288,8 @@ export class Passport {
       error === INVALID_TOKEN_ROTATION ||
       error === NOT_FOUND_ID_TOKEN ||
       error === NOT_FOUND_ACCESS_TOKEN ||
-      error === NOT_FOUND_AUTH_IDENTIIFER;
+      error === NOT_FOUND_AUTH_IDENTIIFER ||
+      error === INVALID_AUTHORIZATION_CODE_FLOW;
 
     return isRefreshTokenError;
   }
@@ -296,13 +298,13 @@ export class Passport {
     try {
       const cache_refresh_token = this.#cacheManager.getRefreshToken();
 
-      if (!cache_refresh_token && !this.#supportWorkerThread) {
+      if (!cache_refresh_token && !this.#supportAuthWorker) {
         throw INVALID_TOKEN_ROTATION;
       }
       if (cache_refresh_token) {
         return { isEnable: true, cache_refresh_token };
       }
-      if (this.#supportWorkerThread) {
+      if (this.#supportAuthWorker) {
         const alive_refresh_token = await checkRefreshTokenAlive(
           'check_refresh_token_alive',
           this.#authWorker as SharedWorker
@@ -367,6 +369,8 @@ export class Passport {
       this.#initializeTransaction();
     }
 
+    console.log('test');
+
     const { url, code_verifier } = await this.#prebuildAuthorizationUrl({
       authorizationOptions: options,
       client_id: clientId,
@@ -389,10 +393,14 @@ export class Passport {
     }
 
     if (queryString.length === 0) {
-      throw new Error(NOT_FOUND_QUERY_PARAMS_ERROR);
+      this.#reconcileAuthorizationCodeFlow(INVALID_AUTHORIZATION_CODE_FLOW, this.#defaultOptions.onLoad);
     }
 
     const { code, error } = parseAuthenticationResult(queryString);
+
+    if (!code) {
+      this.#reconcileAuthorizationCodeFlow(INVALID_AUTHORIZATION_CODE_FLOW, this.#defaultOptions.onLoad);
+    }
 
     const transaction = this.#transactionManager.get();